removed construct methods in controllers and added exception message into Handler.php

+++ app/Http/Controllers/AppointmentController.php
@@ -10,20 +10,12 @@ use Illuminate\Support\Facades\App;
 class AppointmentController extends BaseController
 {
     /**
-     * AppointmentController constructor.
-     * @throws \Illuminate\Auth\Access\AuthorizationException
-     */
-    public function __construct()
-    {
-        $this->authorize('users.personel');
-    }
-
-    /**
      * @return \Illuminate\Http\JsonResponse
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize('users.personel');
         $appointment = Appointment::all();
         return $this->baseMethod($appointment);
     }
@@ -36,6 +28,7 @@ class AppointmentController extends BaseController
      */
     public function store(AppointmentRequest $request)
     {
+        $this->authorize('users.personel');
         $appointment = Appointment::create($request->all());
         return $this->baseMethod($appointment);
     }
@@ -48,6 +41,7 @@ class AppointmentController extends BaseController
      */
     public function show($id)
     {
+        $this->authorize('users.personel');
         $appointment = Appointment::find($id);
         return $this->baseMethod($appointment);
     }
@@ -61,6 +55,7 @@ class AppointmentController extends BaseController
      */
     public function update(AppointmentRequest $request,$id)
     {
+        $this->authorize('users.personel');
         $appointment = Appointment::find($id);
         $appointment->update($request->all());
         return $this->baseMethod($appointment);
@@ -74,6 +69,7 @@ class AppointmentController extends BaseController
      */
     public function destroy($id)
     {
+        $this->authorize('users.personel');
         $appointment = Appointment::find($id)->delete();
         return $this->baseMethod($appointment);
     }
